Create grade.py
Holy moly, because the class's average was super low on the test, the teacher just added a curve for the test grades! 😭

Create a grades.py program that checks whether a grade is above or below 55.

Start by creating a variable called grade and give it a value between 0-100.

Write a if/else statement for the following:

If grade is greater than or equal to 55, then print "You passed."
Else, print "You failed."
After you run the code, change grade's value and rerun it. Do this a few times to make sure it's working as intended.